Bug 1934334 - Use a larger type for a video encoder framerate, clamp the frame rate to positive value. r=media-playback-reviewers,aosmond
There is no reason the frame rate can't be greater than 255.

Differential Revision: https://phabricator.services.mozilla.com/D233268

+++ dom/media/platforms/EncoderConfig.h
@@ -123,7 +123,7 @@ class EncoderConfig final {
   // This constructor is used for video encoders
   EncoderConfig(const CodecType aCodecType, gfx::IntSize aSize,
                 const Usage aUsage, const PixelFormat aPixelFormat,
-                const PixelFormat aSourcePixelFormat, const uint8_t aFramerate,
+                const PixelFormat aSourcePixelFormat, const uint32_t aFramerate,
                 const size_t aKeyframeInterval, const uint32_t aBitrate,
                 const uint32_t aMinBitrate, const uint32_t aMaxBitrate,
                 const BitrateMode aBitrateMode,
@@ -184,7 +184,7 @@ class EncoderConfig final {
   PixelFormat mPixelFormat{};
   PixelFormat mSourcePixelFormat{};
   ScalabilityMode mScalabilityMode{};
-  uint8_t mFramerate{};
+  uint32_t mFramerate{};
   size_t mKeyframeInterval{};
   // Audio-only
   uint32_t mNumberOfChannels{};
